# Improve InterfaceClass __hash__ performance