--- a/CHANGES.rst
+++ b/CHANGES.rst
@@ -17,6 +17,13 @@
   an undocumented private C API function, and helps make some
   instances require less memory. See `PR 154 <https://github.com/zopefoundation/zope.interface/pull/154>`_.
 
+- Performance optimization of ``__hash__`` method on ``InterfaceClass``.
+  The method is called very often (i.e several 100.000 times on a Plone 5.2
+  startup). Because the hash value never changes it can be cached.
+  This improves test performance from 0.614s down to 0.575s (1.07x faster).
+  In a real world Plone case a reindex index came down from 402s to 320s (1.26x faster).
+  See `PR 156 <https://github.com/zopefoundation/zope.interface/pull/156>`_.
+
 
 4.7.1 (2019-11-11)
 ==================