Performance optimization of ``__hash__`` method on ``InterfaceClass``.

The method is called very often (i.e several 100.000 times on a Plone 5.2 startup). 
Even if a single ``hash`` is very fast, this sums up to a real significant part of processing time.
Because the hash value calculated from ``name`` and ``__module__`` never changes (afaik), it can be cached.

I took Py-Spy to analyze the performance.
On reindexing of an index in ZCatalog the ``__hash__`` took ~9 seconds after 10000 samples before changes. 
With a cached hash like in this implementation ``__hash__`` takes ~4.7 seconds.
The whole reindex index process came down from 402s to 320s (1.26x faster).
The test performance ``python setup.py test`` goes from 0.614s down to 0.575s (1.07x faster).