@fschulze 2020-01-24
Looks good to me. If someone does anything which breaks this, they can be expected to adapt to the change.