@jaroel 2020-01-24
I'm pondering if it could be possible that an Interface instance is persisted, and then moved to a different module. How would that affect the hash?
In other words, would it be better/safer to use `self._v_cached_hash`?