@ale-rt 2020-01-25
I added this checkout on Plone5.2 found what could be a problem in some (not all) browser doctests (which are a functional test):
```
...
  File "/var/lib/jenkins/.buildout/eggs/plone.autoform-1.8.1-py3.7.egg/plone/autoform/utils.py", line 287, in processFields
    _process_fieldsets(form, schema, groups, all_fields, prefix, defaultGroup)
  File "/var/lib/jenkins/.buildout/eggs/plone.autoform-1.8.1-py3.7.egg/plone/autoform/utils.py", line 170, in _process_fieldsets
    group.fields += new_fields
  File "/var/lib/jenkins/.buildout/eggs/z3c.form-3.7.0-py3.7.egg/z3c/form/util.py", line 298, in __add__
    return self.__class__(self, other)
  File "/var/lib/jenkins/.buildout/eggs/z3c.form-3.7.0-py3.7.egg/z3c/form/field.py", line 136, in __init__
    raise ValueError("Duplicate name", name)
ValueError: ('Duplicate name', 'IDublinCore.creators')
```
Robot tests (which are also functional) do not suffer from this.

Just mentioning as a note, it might be that plone.autoform needs an update and Plone 5.2 is not really thought to work with zope.interface master.

Abyway for me it is good to go :)