@jamadden 2020-01-27
> I'm pondering if it could be possible that an Interface instance is persisted, and then moved to a different module. How would that affect the hash?
In other words, would it be better/safer to use self._v_cached_hash?

`InterfaceClass` defines `__reduce__` which is used in preference to the default pickle behaviour of saving an instance's `__dict__`. `__reduce__` simply returns the name of the interface (which is automatically combined with its `__module__` by the pickle machinery) so unpickling turns into a global lookup of the object with the same name and module. Also, `_v_` attributes are only special to subclasses of `persistent.Persistent`, which `InterfaceClass` isn't.

So this would matter only to a a subclass of `InterfaceClass` that overrides `__reduce__`, potentially by subclassing Persistent.

On current master, it's not possible to subclass both `Persistent` and `InterfaceClass`:
```pycon
>>> from zope import interface
>>> from persistent import Persistent
>>> class PersistentInterfaceClass(Persistent, 
          interface.interface.InterfaceClass):
…        pass 
TypeError: multiple bases have instance lay-out conflict
```

On current released versions, it's not possible to pickle such an object because there are things in `__dict__` that can't be pickled:
```pycon
>>> from zope import interface
>>> from persistent import Persistent
>>> class PersistentInterfaceClass(Persistent,
           interface.interface.InterfaceClass):
...     pass
...
>>> class IP(metaclass=PersistentInterfaceClass): pass
...
>>> IP
<__main__.PersistentInterfaceClass object at 0x10f5c59d0>
>>> type(IP)
<class '__main__.PersistentInterfaceClass'>
>>> import pickle
>>> pickle.dumps(IP)
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
AttributeError: Can't pickle local object 'WeakKeyDictionary.__init__.<locals>.remove'
```

If you switch the order of inheritance around you can pickle it, but it goes back to pickling by name using `__reduce__` and ignoring the persistence machinery. ZODB fails to pickle `IP` and *also* fails to pickle the reversed order of inheritance because there's no `__getstate__` defined. 

So it would take subclassing `Persistent` and `InterfaceClass` in the right order, and also defining `__getstate__` in this subclass to attempt to pickle an interface in this way. Is anyone aware of something that does that, or that has a use case to do that?